pyproject.toml
[project]
name = "TextEvolve"
version = "0.1.0"
description = ""
authors = ["Nick Ryan <nickcdryan@gmail.com>"]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.72.0",
    "sift-stack-py>=0.5.0",
    "google-genai>=1.10.0",
    "networkx>=3.4.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "requests>=2.32.0",
    "docker>=7.0.0",
    "google>=3.0.0",
    "PyYAML>=6.0.0",
]
